internal/api: implement dns handler
This commit adds the `api` package that will contain the various API
endpoints the server provides. To start, we've got DNS handling where
questions are checked against the allow & block lists. If a question
appears in the allow list, or not in either list, it is upstreamed to
a configured set of servers.

The api attempts each configured upstream until it gets a successful
result.

Signed-off-by: David Bond <davidsbond93@gmail.com>