internal/api: use weighted slice for upstream selection
This commit modifies the upstream selection process to use a weighted
slice (via my experimental packages) to modify the range order of
upstreams by their round-trip time.

This should make it so that we always pick the fastest upstream we're
aware of. All of the weightings are initially set as zero and range
occurs in ascending order for lowest round-trip time.

This means that for the first N dns queries (where N is the number of
upstreams), we'll pick one we have yet to use which allows us to set
up the initial round-trip times/weights.

Signed-off-by: David Bond <davidsbond93@gmail.com>