internal/{server,handler}: initial support for dns-over-https
This commit renames the api package to "handler" and adds a new
ServeHTTP method that implements the expected POST/GET http handling
code for dns-over-https requests.

A lot of refactoring has gone on to make parts reusable as well as tweaks
and fixes for better fulfilment of the various DNS RFCs.

Signed-off-by: David Bond <davidsbond93@gmail.com>