internal/handler: more RFC compliance
This commit adds appropriate error handling when dealing with EDNS0 stuff.
Primarily version checking and rejecting anything that does not have a
version of zero.

On top of this I (with the help of ChatGPT) have attempted to annotate
code paths with the RFCs they relate to and listed all the RFCs at the
top of the file.

Signed-off-by: David Bond <davidsbond93@gmail.com>