internal/{cache,handler,server}: add DNS query caching
This commit adds caching for DNS queries using ristretto as the underlying
cache implementation.

The cache is RFC compliant in how it generate cache keys and modifies TTLs
on query to provide the TTL of the remaining time in the cache rather than
the original value given by the upstream.

A bit of tidying up has been included as the Handler type now has a lot
of params. These are now all contained within the handler.Config type.

When using the default configuration the server will enable caching.

Signed-off-by: David Bond <davidsbond93@gmail.com>